Issue #2: culling old podcasts
It looks like commit 9dfc140 borked changing the enclosure url to filesystem
path by either stripping off and extra / or by not matching because there
were not two slashes. Behavior depends on the values of --baseurl and --basedir
and could (depending on values) behave correctly, not match at all, or build
an invalid path.

I am taking out the trailing / on the match and putting an extra in on the
replacement under this thought process:
    if --baseurl has a trailing /
        we need to add a trailing slash to separate --basedir from --podcast
    else
        an extra one will not hurt (in my experience on *nix and *dose).
    if --basedir already has a trailing /
        an extra one will not hurt (see above)